refactor: move logic to calendar-graph.service.ts
